% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{comparisonplot}
\alias{comparisonplot}
\title{Qualitative plot of comparison for two word clouds}
\usage{
comparisonplot(freq, counts, qvals, colors = c("red", "blue", "black"),
  size.limits = 50:1/20, xlim = c(-0.5, 1), ylim = c(-0.5, 1))
}
\arguments{
\item{freq}{A numeric matrix with two columns. The first column indicates the
frequency of a term in the first group, and the second column indicates
frequency of a term in the second group. The rows describe terms. Rownames
must be set to the terms.}

\item{counts}{A numeric matrix in the same format as \code{freq} except the entries
represent count values for each of the groups.}

\item{qvals}{A vector of q-values for each of the terms. The names for the vector
must be the terms. Q-values of 0 are rounded to 0.00001. Note that there
must be at least two distinct q-values (i.e., they cannot all be 1).}

\item{colors}{A character vector of length three. The first element represents the color
that the term should be plotted in if the count value is greater in the
first group compared to the second group. The second element represents
the color for terms with greater counts in the second group compared to
the first group. The third element is the color for terms that have equivalent
counts in both groups.Defaults to \code{c('red','blue','black')}}

\item{size.limits}{Bounds for the sizes of the words. Must be provided in the format max:min.
The sizes are interpreted by \code{cex}, so they are scaled. Defaults to \code{50:1/20}.}

\item{xlim}{A numeric vector of length two with the lower and upper bounds for the x-axis of the graph.
Defaults to \code{c(-0.5,1)}.}

\item{ylim}{A numeric vector of length two with the lower and upper bounds for the y-axis of the graph.
            Defaults to \code{c(-0.5,1)}.}

\item{direction}{A vector of numeric values indicating whether the term belongs to
group 1 or 2. The names of the vector must be the terms.}
}
\description{
Main user interface for qualitative analysis of two word clouds.
Creates a plot with frequencies of each group on each axis. The size of each
term is inversely related to the q-values provided. The color of each term
indicates which group had the greater number of counts for the term occurrence.

Note that all parameters must describe the same number of terms.
}
\examples{
group1 <- c('head','toe','hand',rep('knee',4))
group2 <- c(rep('toe',3),'hand',rep('head',2))

stats <- wordcloudstats(group1, group2, names=c('Group 1','Group 2'))
qval <- runif(nrow(stats$frequency),0.0,1.0)
names(qval) <- rownames(stats$outputs)
comparisonplot(stats$frequency, stats$counts, qval, colors=c('blue','green','red'))
}
\seealso{
\code{\link[wordcloud]{wordlayout}}
}

