% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{wordcloudanalysis-package}
\alias{wordcloudanalysis-package}
\title{Statistical tests to compare word clouds}
\description{
\code{\link{wordcloudstats}} is the main user interface to conduct a quantitative analysis of
the significance of the observed differences between the two word clouds.
}

